made a wavesurfer component specifically for listening (student testing experience)
